改为  cmd 模块
