+++ .gitignore
@@ -1,4 +1,23 @@
+*.iml
+.idea/
+.ipr
+.iws
+*~
+~*
+*.diff
+*.patch
+*.bak
 .DS_Store
-nuget/content
-*.nupkg
-tests/phantomjs
+Thumbs.db
+.project
+.*proj
+.svn/
+*.swp
+*.swo
+*.pyc
+*.pyo
+build
+node_modules
+_site
+sea-modules
+.cache
