+++ .travis.yml
@@ -1,0 +1,19 @@
+language: node_js
+
+node_js:
+  - 0.10
+
+install:
+  - npm install mocha-browser nico
+
+before_script:
+  - git clone git://github.com/aralejs/nico-arale.git _theme
+  - node_modules/.bin/nico build --theme _theme -C _theme/nico.js
+
+script:
+  - node_modules/.bin/mocha-browser _site/tests/runner.html -S
+
+after_success:
+  - npm install jscoverage coveralls
+  - node_modules/.bin/jscoverage --encoding=utf8 src _site/src-cov
+  - node_modules/.bin/mocha-browser _site/tests/runner.html?cov -S -R lcov | node_modules/.bin/coveralls
