+++ README.md
@@ -1,41 +1,11 @@
-# 说明
+# bootstrap3-grid
 
-源代码来自 [https://github.com/JamesRandall/Simple-Paging-Grid](https://github.com/JamesRandall/Simple-Paging-Grid "https://github.com/JamesRandall/Simple-Paging-Grid")
+---
+源码来自[Simple-Paging-Grid](https://github.com/JamesRandall/Simple-Paging-Grid)，把它改为了CMD模块的
 
-因不想更改后台所给的数据接口格式，主要是对 ajax 返回的数据进行更改，且只更改0.6.0.0版本
+---
 
-## 改动
-
-
-	/**
-     * currentPage 改为 data.listData , totalRows 改为 data.totalCount, 并添加rowIndex
-     *
-     * @param sourceData
-     * @private
-     */
-    _parseSourceData: function(sourceData) {
-        this._sourceData = sourceData;
-        if ($.isArray(sourceData)) {
-            this._pageData = sourceData;
-            this._numberOfRows = null;
-        } else if ($.isPlainObject(sourceData)) {
-            // 给数据添加 rowIndex
-            $.each(sourceData.data.listData, function(index) {
-                sourceData.data.listData[index]['rowIndex'] = index + 1;
-            })
-            this._pageData = sourceData.data.listData;
-            this._numberOfRows = sourceData.data.totalCount;
-        }
-        else if (sourceData === null || sourceData === undefined) {
-            this._pageData = [];
-            this._numberOfRows= 0;
-        }
-        this._deferredCellTemplateCompilation();
-    }
-
-## 例子
-
-官方地址
+## 官方例子
 
 1. Basic grid
 <http://www.accidentalfish.com/simplePagingGrid/examples/basicGrid/index.html>
@@ -55,51 +25,21 @@
 6. Refreshing a data array
 <http://www.accidentalfish.com/simplePagingGrid/examples/dataArrayRefresh/index.html>
 
-## 基本用法
+## Usage
 
-引入样式文件
+It is very easy to use this module.
 
-	<link rel="stylesheet" href="/css/simplePagingGrid-0.6.0.0.css">
+````html
+<div class="zeus-bootstrap3-grid">
+</div>
+````
 
-引入脚本文件
+```javascript
+seajs.use('bootstrap3-grid', function(bootstrap3Grid) {
 
-	<script src="/script/jquery-1.8.3.min.js" type="text/javascript">
-    <script src="/script/handlebars-1.0.rc.1.js" type="text/javascript">
-    <script src="/script/simplePagingGrid-0.6.0.0.js" type="text/javascript">
+});
+```
 
-必须先引入 jQuery 和 Handlebars :
+## Api
 
-<http://jquery.com/>
-
-<http://handlebarsjs.com/>
-
-在 html 里创建一 div
-
-    <div id="exampleGrid"/>
-
-In your document ready event handler configure your grid
-
-    $(document).ready(function() {
-        $("#exampleGrid").simplePagingGrid({
-            data: [
-                { "Name": "Pineapple", "Price": 1.50, "Quantity": 2 },
-                { "Name": "Banana", "Price": 0.30, "Quantity": 5 }]
-        });
-    });
-
-The grid will look at your data array and create column headers and keys from the first item. Normally you want a little more sugar and using the columnNames and columnKeys options will let you do that
-
-    $(document).ready(function() {
-        $("#exampleGrid").simplePagingGrid({
-            columnNames: ["Name", "Price ($)", "Quantity"],
-            columnKeys: ["Name", "Price", "Quantity"],
-            data: [
-                { "Name": "Pineapple", "Price": 1.50, "Quantity": 2 },
-                { "Name": "Banana", "Price": 0.30, "Quantity": 5 }]
-        });
-    });
-
-更多例子见官方
-
-
-
+Here is some details.
