+++ examples/cellTemplates/index.html
@@ -9,35 +9,37 @@
 
     <!-- Le styles -->
     <link href="css/bootstrap.min.css" rel="stylesheet">
-    <script type="text/javascript" src="script/jquery-1.8.3.min.js"></script>
-	<script type="text/javascript" src="script/handlebars-1.0.rc.1.js"></script>
-    <script type="text/javascript" src="../../script/simplePagingGrid-0.6.0.0.js"></script>
+    <script type="text/javascript" src="../../sea-modules/seajs/seajs/2.1.1/sea-debug.js"></script>
+    <script src="../sea-config.js"></script>
 	<script type="text/javascript">
-		$(document).ready(function() {
-		        var rows = [];
-		        var rowNumber;
-		        var numberOfRows = 355;
-		        for(rowNumber = 0; rowNumber < numberOfRows; rowNumber++) {
-		            rows.push({
-		                "OrderLineID": rowNumber,
-		                "Name": "Product " + rowNumber,
-		                "Price": Math.floor(Math.random() * 100),
-		                "Quantity": Math.floor(Math.random() * 100)
-		            });
-		        }
-		        
-              $("#exampleGrid").simplePagingGrid({
-                cellTemplates: [null, "${{Price}}", null, "{{../message}}"],
-                columnNames: ["Name", "Price", "Quantity", "Additional Payload"],
-                columnKeys: ["Name", "Price", "Quantity", null],
-                columnWidths: ["25%", "25%", "25%", "25%"],
-                data: {
-                    currentPage: rows,
-                    totalRows: rows.length,
-                    message: "Additional payload"
+        seajs.use(['$', '../demo'], function($) {
+            $(document).ready(function() {
+                var rows = [];
+                var rowNumber;
+                var numberOfRows = 355;
+                for(rowNumber = 0; rowNumber < numberOfRows; rowNumber++) {
+                    rows.push({
+                        "OrderLineID": rowNumber,
+                        "Name": "Product " + rowNumber,
+                        "Price": Math.floor(Math.random() * 100),
+                        "Quantity": Math.floor(Math.random() * 100)
+                    });
+                }
+
+                $("#exampleGrid").simplePagingGrid({
+                    cellTemplates: [null, "${{Price}}", null, "{{../message}}"],
+                    columnNames: ["Name", "Price", "Quantity", "Additional Payload"],
+                    columnKeys: ["Name", "Price", "Quantity", null],
+                    columnWidths: ["25%", "25%", "25%", "25%"],
+                    data: {
+                        currentPage: rows,
+                        totalRows: rows.length,
+                        message: "Additional payload"
                     }
-		        });
-		    });
+                });
+            });
+        })
+
 	</script>
 </head>
 <body>
