+++ examples/demo.js
@@ -1,0 +1,3 @@
+define(function(require) {
+    require('../src/bootstrap3-grid');
+})
