+++ examples/hyperlinks/index.html
@@ -9,20 +9,21 @@
 
     <!-- Le styles -->
     <link href="css/bootstrap.min.css" rel="stylesheet">
-    <script type="text/javascript" src="script/jquery-1.8.3.min.js"></script>
-    <script type="text/javascript" src="script/handlebars-1.0.rc.1.js"></script>
-    <script type="text/javascript" src="script/data.js"></script>
-    <script type="text/javascript" src="../../script/simplePagingGrid-0.6.0.0.js"></script>
+    <script type="text/javascript" src="../../sea-modules/seajs/seajs/2.1.1/sea-debug.js"></script>
+    <script src="../sea-config.js"></script>
 	<script type="text/javascript">
-		$(document).ready(function() {
-		        $("#exampleGrid").simplePagingGrid({
-		            columnNames: ["Name", "Price ($)", "Quantity"],
-		            columnKeys: ["Name", "Price", "Quantity"],
-		            columnWidths: ["50%", "25%", "25%"],
-		            cellTemplates:["<a href='orderLineView.html?id={{OrderLineID}}'>{{Name}}</a>", null, "{{Quantity}} units"],
-		            data: getData()
-		        });
-		    });
+        seajs.use(['$','../demo'], function($) {
+            $(document).ready(function() {
+                $("#exampleGrid").simplePagingGrid({
+                    columnNames: ["Name", "Price ($)", "Quantity"],
+                    columnKeys: ["Name", "Price", "Quantity"],
+                    columnWidths: ["50%", "25%", "25%"],
+                    cellTemplates:["<a href='orderLineView.html?id={{OrderLineID}}'>{{Name}}</a>", null, "{{Quantity}} units"],
+                    data: getData()
+                });
+            });
+        })
+
 	</script>
 </head>
 <body>
