+++ examples/index.md
@@ -1,0 +1,11 @@
+# Demo
+
+---
+
+## Normal usage
+
+````javascript
+seajs.use('bootstrap3-grid', function(bootstrap3Grid) {
+
+});
+````
