+++ examples/sea-config.js
@@ -1,0 +1,7 @@
+seajs.config({
+    base: '../../sea-modules',
+    alias: {
+        jquery: 'jquery/jquery/1.10.1/jquery',
+        $: 'jquery/jquery/1.10.1/jquery'
+    }
+});
