--- nuget/SimplePagingGrid.nuspec
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>SimplePagingGrid</id>
-    <version>0.6.0.0</version>
-    <title>Simple Paging Grid</title>
-    <authors>James Randall</authors>
-    <owners>James Randall</owners>
-    <licenseUrl>https://raw.github.com/JamesRandall/Simple-Paging-Grid/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/JamesRandall/Simple-Paging-Grid</projectUrl>
-    <iconUrl>http://www.accidentalfish.com/simplePagingGridIcon.png</iconUrl>
-    <requireLicenseAcceptance>true</requireLicenseAcceptance>
-    <description>Bootstrap based JavaScript grid built as a jQuery plugin and making use of the Handlebars templating library. Super simple to use and very customizable.</description>
-    <releaseNotes>Now works with Bootstrap 3 by default (to use Bootstrap 2 set the bootstrapVersion property to 2), removed the need for the grids own CSS file, addressed GitHub logged issues, updated dataFunction to support a callback approach, resolved issue with emptyTemplate and null results, added support for URL updating to enable back and forward button behaviour.</releaseNotes>
-    <copyright>Copyright 2013 Accidental Fish Ltd.</copyright>
-    <tags>grid datagrid jquery plugins widgets javascript handlebars bootstrap simple json mvc asp.net</tags>
-    <dependencies>
-      <dependency id="Handlebars.js" version="1.0.7" />
-      <dependency id="jQuery" version="1.8.3" />
-    </dependencies>
-  </metadata>
-</package>
