+++ src/bootstrap3-grid.css
@@ -1,4 +1,8 @@
-﻿ul.sort {
+div.sort-container {
+    cursor: pointer;
+}
+
+ul.sort {
     list-style: none;
     margin: 0px;
     padding: 0px;
@@ -37,3 +41,17 @@ li.sort-descending-active
 {
     border-color: #aaa transparent transparent transparent;
 }
+
+div.loading
+{
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    z-index: 999;
+    border: 5px;
+    background-color: white;
+    background-image: url('load.gif');
+    background-repeat: no-repeat;
+    background-position: center;
+}
+
