+++ tests/bootstrap3-grid-spec.js
@@ -1,0 +1,12 @@
+define(function(require) {
+
+  var bootstrap3Grid = require('../src/bootstrap3-grid');
+
+  describe('bootstrap3-grid', function() {
+
+    it('normal usage', function() {
+
+    });
+  });
+
+});
